requirements.txt
absl-py==0.5.0
appnope==0.1.0
astor==0.7.1
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==2.1.4
Click==7.0
configparser==3.5.0
decorator==4.3.0
defusedxml==0.5.0
Django==1.11.15
entrypoints==0.2.3
enum34==1.1.6
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.2.0
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
html5lib==1.0.1
image==1.5.25
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.2
Keras-Applications==1.0.5
Keras-Preprocessing==1.0.3
Markdown==3.0.1
MarkupSafe==1.0
mistune==0.8.3
mock==2.0.0
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
np-utils==0.5.5.2
numpy==1.15.2
pandocfilters==1.4.2
pathlib2==2.3.2
pbr==4.2.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.2.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
scandir==1.9.0
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.12
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.1
Theano==1.0.3
tornado==5.1.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
